+++ widget/gotop/src/gotop.js
@@ -24,7 +24,7 @@ function goTopInit() {
 
   checkPosition();
 
-  $win.on('scroll.gotop.amui', $.AMUI.utils.debounce(checkPosition, 100));
+  $win.on('scroll.gotop.amui', UI.utils.debounce(checkPosition, 100));
 
   $goTop.data('init', true);
 }